Rename command handling module
