+++ lib/commandHandler.js
