+++ lib/iotagent-mqtt.js
@@ -26,7 +26,7 @@
 var iotAgentLib = require('iotagent-node-lib'),
     mqtt = require('mqtt'),
     dateFormat = require('dateformat'),
-    commandHandler = require('./commandHandlers'),
+    commandHandler = require('./commandHandler'),
     logger = require('logops'),
     async = require('async'),
     errors = require('./errors'),
