+++ test/deviceProvisioning/provisionConfigurationCommand1.json
